pack default installation tests in one suite
